+++ mp-main/src/main/java/ru/pavbatol/myplace/MainService.java
@@ -1,12 +1,12 @@
-package ru.pavbatol.marketplace;
+package ru.pavbatol.myplace;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MarketplaceStatsService {
+public class MainService {
     public static void main(String[] args) {
-        SpringApplication.run(MarketplaceStatsService.class, args);
+        SpringApplication.run(MainService.class, args);
     }
 
 }
