+++ mp-stats/src/main/java/ru/pavbatol/myplace/StatsService.java
@@ -1,12 +1,12 @@
-package ru.pavbatol.marketplace;
+package ru.pavbatol.myplace;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MarketplaceGatewayService {
+public class StatsService {
     public static void main(String[] args) {
-        SpringApplication.run(MarketplaceGatewayService.class, args);
+        SpringApplication.run(StatsService.class, args);
     }
 
 }
